Warning police: eliminate all defaulting within stage1 Defaulting makes compilation of multiple modules more complicated (re: #1405) Although it was all locally within functions, not because of the module monomorphism-restriction... but it's better to be clear what's meant, anyway. I changed some that were defaulting to Integer, to explicit Int, where Int seemed appropriate rather than Integer.

+++ compiler/main/BreakArray.hs
@@ -26,8 +26,8 @@ import Constants
 data BreakArray = BA (MutableByteArray# RealWorld)
 
 breakOff, breakOn :: Word
-breakOn  = fromIntegral 1
-breakOff = fromIntegral 0
+breakOn  = 1
+breakOff = 0
 
 -- XXX crude
 showBreakArray :: BreakArray -> IO ()
